Доработка задания module5-task1
