+++ js/main.js
@@ -1,1 +1,4 @@
 import './data.js';
+import { createPhotos } from './data.js';
+const somePhotos = createPhotos(25);
+console.log(somePhotos);
